requirements.txt
ipykernel
polygon-api-client
sqlalchemy
psycopg2-binary
pandas
python-dotenv
tqdm
matplotlib
torch
tensorboard
torch-tb-profiler
scikit-learn
mlflow
xgboost
catboost
lightgbm
optuna
optuna-integration[mlflow]